Remove the root (Component) dialog.
